move admin-only pages to admin module
