+++ src/app/admin/admin-routing.module.ts
@@ -1,7 +1,41 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-const routes: Routes = [];
+import { AuthGuard } from 'app/core';
+import { PostDetailResolver, DraftListResolver } from 'app/blogging-core';
+import { CanDeactivateDirtyPost } from 'app/blogging-core';
+
+import { PostCreateComponent } from './post-create/post-create.component';
+import { PostEditComponent } from './post-edit/post-edit.component';
+import { DraftsComponent } from './drafts/drafts.component';
+
+const routes: Routes = [
+  {
+    path: '',
+    canActivate: [AuthGuard],
+    children: [
+      {
+        path: 'create',
+        component: PostCreateComponent,
+        canDeactivate: [CanDeactivateDirtyPost],
+        data: { pageId: 'postCreate', title: 'Create post' },
+      },
+      {
+        path: 'drafts',
+        component: DraftsComponent,
+        resolve: { posts: DraftListResolver },
+        data: { pageId: 'drafts', title: 'Drafts' },
+      },
+      {
+        path: ':pk/edit',
+        component: PostEditComponent,
+        canDeactivate: [CanDeactivateDirtyPost],
+        resolve: { post: PostDetailResolver },
+        data: { pageId: 'postEdit' },
+      },
+    ],
+  },
+];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
