+++ src/app/admin/drafts/drafts.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { Post } from '../core';
+import { Post } from 'app/blogging-core';
 
 @Component({
   selector: 'app-drafts',
