+++ src/app/admin/post-create/post-create.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
 import { tap } from 'rxjs/operators';
-import { Post, PostPayload, PostService, DirtyComponent } from '../core';
+import { Post, PostPayload, PostService, DirtyComponent } from 'app/blogging-core';
 
 
 @Component({
@@ -15,7 +15,8 @@ export class PostCreateComponent implements DirtyComponent {
 
   constructor(
     private router: Router,
-    private postService: PostService) { }
+    private postService: PostService,
+  ) { }
 
   onSubmit(payload: PostPayload) {
     this.postService.create(payload).pipe(
