+++ src/app/admin/post-edit/post-edit.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
 import { tap } from 'rxjs/operators';
-import { Post, PostPayload, PostService, DirtyComponent } from '../core';
+import { Post, PostPayload, PostService, DirtyComponent } from 'app/blogging-core';
 
 
 @Component({
