+++ src/app/blog/post-footer/post-footer.component.ts
@@ -1,15 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-post-footer',
   templateUrl: './post-footer.component.html',
   styleUrls: ['./post-footer.component.scss']
 })
-export class PostFooterComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
+export class PostFooterComponent { }
